// src/app/interfaces/training-position.ts
export interface TrainingPosition {

    position_id: number;
    position_description: string;
    position_posting_date: string;
    position_visibility: boolean;
    position_status: string;
    position_free_places: number;
    position_work_type: string;
    position_starting_date: string;
    position_duration: number;
    position_requirements: string;
    position_image : string; 
    regions: string[]; 
    notifications: any[]; 
    company :string ; 
            
}